ファイル名のピリオド削除処理をメソッド抽出。#6
